+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added support for anisotropic filtering to xc3_wgpu.
 * Added support for `monolib/shader` textures to xc3_wgpu.
+* Added options to set the mipmaps and quality when generating compressed files to xc3_tex.
 
 ### Changed
 * Improved the readability and quality of displayed error messages for xc3_viewer, xc3_gltf, and xc3_tex.